test: add Log in with a new user account test
